test(flusher): merge parity tests into plugin test files
Remove shared flusherparity fixtures; reuse each flusher's
makeTestLogGroupList sample data for v1 Flush vs v2 Export checks.
