+++ plugins/flusher/statistics/flusher_statistics_test.go
@@ -1,4 +1,4 @@
-// Copyright 2021 iLogtail Authors
+// Copyright 2026 iLogtail Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,49 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kafka
+package statistics
 
 import (
-	"encoding/json"
 	"strconv"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/alibaba/ilogtail/pkg/models"
 	"github.com/alibaba/ilogtail/pkg/protocol"
-	"github.com/alibaba/ilogtail/plugins/test"
 	"github.com/alibaba/ilogtail/plugins/test/mock"
 )
 
-// Invalid Test
-func InvalidTestConnectAndWrite(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration test in short mode")
+func createLogByFields(fields map[string]string) *protocol.Log {
+	var slsLog protocol.Log
+	for key, val := range fields {
+		slsLog.Contents = append(slsLog.Contents, &protocol.Log_Content{Key: key, Value: val})
 	}
-
-	brokers := []string{"172.17.0.2:9092"}
-	k := &FlusherKafka{
-		Brokers:      brokers,
-		Topic:        "Test",
-		SASLUsername: "",
-		SASLPassword: "",
-	}
-
-	// Verify that we can connect to the Kafka broker
-	lctx := mock.NewEmptyContext("p", "l", "c")
-	err := k.Init(lctx)
-	require.NoError(t, err)
-
-	// Verify that we can successfully write data to the kafka broker
-	lgl := makeTestLogGroupList()
-	err = k.Flush("projectName", "logstoreName", "configName", lgl.GetLogGroupList())
-	require.NoError(t, err)
-	_ = k.Stop()
+	protocol.SetLogTime(&slsLog, uint32(time.Now().Unix()))
+	return &slsLog
 }
 
 func makeTestLogGroupList() *protocol.LogGroupList {
-	f := map[string]string{}
+	fields := map[string]string{}
 	lgl := &protocol.LogGroupList{
 		LogGroupList: make([]*protocol.LogGroup, 0, 10),
 	}
@@ -64,10 +46,9 @@ func makeTestLogGroupList() *protocol.LogGroupList {
 			Source: "",
 		}
 		for j := 1; j <= 10; j++ {
-			f["group"] = strconv.Itoa(i)
-			f["message"] = "The message: " + strconv.Itoa(j)
-			l := test.CreateLogByFields(f)
-			lg.Logs = append(lg.Logs, l)
+			fields["group"] = strconv.Itoa(i)
+			fields["message"] = "The message: " + strconv.Itoa(j)
+			lg.Logs = append(lg.Logs, createLogByFields(fields))
 		}
 		lgl.LogGroupList = append(lgl.LogGroupList, lg)
 	}
@@ -105,31 +86,19 @@ func makePipelineGroupEventsFromLogGroups(v1Groups []*protocol.LogGroup) []*mode
 	return v2Groups
 }
 
-func TestFlusherKafka_FlushExportParity(t *testing.T) {
+func TestFlusherStatistics_FlushExportParity(t *testing.T) {
 	v1Groups, v2Groups := makeParityLogGroups()
 
-	var flushPayloads, exportPayloads []string
-	newCapture := func(dst *[]string) FlusherFunc {
-		return func(_ string, _ string, _ string, logGroupList []*protocol.LogGroup) error {
-			for _, lg := range logGroupList {
-				for _, log := range lg.Logs {
-					buf, err := json.Marshal(log)
-					require.NoError(t, err)
-					*dst = append(*dst, string(buf))
-				}
-			}
-			return nil
-		}
-	}
-
-	flushFlusher := &FlusherKafka{flusher: newCapture(&flushPayloads)}
+	flushFlusher := &FlusherStatistics{RateIntervalMs: 1_000_000, GeneratePB: true, SleepMsPerLogGroup: 0}
+	require.NoError(t, flushFlusher.Init(mock.NewEmptyContext("p", "l", "c")))
 	require.NoError(t, flushFlusher.Flush("", "", "", v1Groups))
+	flushGroups := flushFlusher.loggroupRateCounter.Rate()
 
-	exportFlusher := &FlusherKafka{flusher: newCapture(&exportPayloads)}
+	exportFlusher := &FlusherStatistics{RateIntervalMs: 1_000_000, GeneratePB: true, SleepMsPerLogGroup: 0}
+	require.NoError(t, exportFlusher.Init(mock.NewEmptyContext("p", "l", "c")))
 	require.NoError(t, exportFlusher.Export(v2Groups, nil))
+	exportGroups := exportFlusher.loggroupRateCounter.Rate()
 
-	require.Equal(t, len(flushPayloads), len(exportPayloads))
-	for i := range flushPayloads {
-		require.JSONEq(t, flushPayloads[i], exportPayloads[i])
-	}
+	require.Equal(t, flushGroups, exportGroups)
+	require.Equal(t, flushFlusher.logRateCounter.Rate(), exportFlusher.logRateCounter.Rate())
 }
